Refactor backend codebase for improved structure and maintainability
- Rename 'src/celery' to 'src/queue' for better naming
- Move database-related files to 'src/database'
- Reorganize CLI commands into 'src/cli' with submodules
- Consolidate models into top-level 'models.py'
- Update all import paths to match new structure
- Fix formatting and import issues in various files
- Remove unused 'services' target from Makefile"
