--- backend/src/batch/__init__.py
@@ -1,8 +1,0 @@
-"""Batch processing module for UI element detection."""
-
-from src.batch.processor import BatchProcessor, auto_predict_images
-
-__all__ = [
-    "BatchProcessor", 
-    "auto_predict_images"
-]
