--- backend/src/celery/__init__.py
@@ -1,18 +1,0 @@
-"""Celery task queue components."""
-
-from src.celery.app import celery_app
-from src.celery.models import Job, JobStatus
-from src.celery.tasks import (
-    process_image_task,
-    check_queue_size_task,
-    cleanup_old_jobs_task
-)
-
-__all__ = [
-    "celery_app",
-    "Job",
-    "JobStatus",
-    "process_image_task", 
-    "check_queue_size_task",
-    "cleanup_old_jobs_task"
-]
