+++ backend/src/cli/evaluate/__init__.py
@@ -1,0 +1,1 @@
+"""CLI tools for UI annotation evaluation."""
