+++ backend/src/cli/main.py
@@ -7,23 +7,25 @@ from pathlib import Path
 
 import click
 
-# Import evaluation commands
-from src.evaluate.detection_evaluator import evaluate_detection_performance
-from src.evaluate.formatter import print_evaluation_results, format_results_as_json
-
 # Import batch processor
-from src.batch.processor import auto_predict_images
+from src.cli.batch_predict.processor import auto_predict_images
+
+# Import evaluation commands
+from src.cli.evaluate.detection_evaluator import evaluate_detection_performance
+from src.cli.evaluate.formatter import format_results_as_json, print_evaluation_results
 
 # Import settings to get default model
 from src.settings import config
+from src.constants import DEFAULT_PREDICTIONS_DIR, DEFAULT_GROUND_TRUTH_DIR, DEFAULT_IMAGE_DIR
 
-DEFAULT_PREDICTIONS_DIR = Path(__file__).parent.parent / "dataset" / "labels" /"predictions"
-DEFAULT_GROUND_TRUTH_DIR = Path(__file__).parent.parent / "dataset" / "labels" / "ground_truth"
-DEFAULT_IMAGE_DIR = Path(__file__).parent.parent / "dataset" / "images"
+# Convert relative paths to absolute paths based on project root
+DEFAULT_PREDICTIONS_DIR = Path(__file__).parent.parent.parent / DEFAULT_PREDICTIONS_DIR
+DEFAULT_GROUND_TRUTH_DIR = Path(__file__).parent.parent.parent / DEFAULT_GROUND_TRUTH_DIR
+DEFAULT_IMAGE_DIR = Path(__file__).parent.parent.parent / DEFAULT_IMAGE_DIR
 
 class CustomGroup(click.Group):
     """Custom Group class that formats commands with multi-line help."""
-    
+
     def format_commands(self, ctx, formatter):
         """Format commands with multi-line support."""
         commands = []
@@ -36,7 +38,7 @@ class CustomGroup(click.Group):
             # Get the short help, preserving newlines
             help_text = cmd.get_short_help_str(limit=500) or ""
             commands.append((subcommand, help_text))
-        
+
         if commands:
             with formatter.section('Commands'):
                 # Write each command with proper multi-line formatting
@@ -52,7 +54,7 @@ class CustomGroup(click.Group):
 @click.version_option(version='0.2.0', prog_name='cli')
 def cli():
     """UI Annotation CLI Tools.
-    
+
     A collection of tools for UI element detection and evaluation.
     """
     pass
@@ -70,19 +72,19 @@ def evaluate(ground_truth_dir: Path, predictions_dir: Path, iou_threshold: float
         ground_truth_dir = DEFAULT_GROUND_TRUTH_DIR
     if predictions_dir is None:
         predictions_dir = DEFAULT_PREDICTIONS_DIR
-        
+
     # Check if directories exist when using defaults
     if not ground_truth_dir.exists():
         raise click.ClickException(f"Ground truth directory not found: {ground_truth_dir}")
     if not predictions_dir.exists():
         raise click.ClickException(f"Predictions directory not found: {predictions_dir}")
-    
+
     # Run evaluation
     tag_metrics = evaluate_detection_performance(ground_truth_dir, predictions_dir, iou_threshold)
-    
+
     # Print results
     print_evaluation_results(tag_metrics, show_errors=show_errors)
-    
+
     # Save results if output file specified
     if output:
         params = {
@@ -90,13 +92,13 @@ def evaluate(ground_truth_dir: Path, predictions_dir: Path, iou_threshold: float
             'predictions_dir': str(predictions_dir),
             'iou_threshold': iou_threshold
         }
-        
+
         results = format_results_as_json(tag_metrics, params)
-        
+
         # Write to file
         with open(output, 'w') as f:
             json.dump(results, f, indent=2)
-        
+
         click.echo(f"\nResults saved to: {output}")
 
 
@@ -113,30 +115,30 @@ def batch_predict(image_dir: str, output_dir: str, model: str, concurrent: int, 
             raise click.ClickException("No models configured in model_config.yaml")
         model = next(iter(config.models.keys()))
         click.echo(f"Using default model: {model}")
-    
+
     # Validate model exists
     if model not in config.models:
         available_models = ", ".join(config.models.keys())
         raise click.ClickException(f"Model '{model}' not found. Available models: {available_models}")
-    
+
     if image_dir:
         image_path = Path(image_dir)
     else:
         image_path = DEFAULT_IMAGE_DIR
         click.echo(f"Using default image directory: {image_path}")
-    
+
     if output_dir:
         output_path = Path(output_dir)
     else:
         output_path = DEFAULT_PREDICTIONS_DIR
         click.echo(f"Using default output directory: {output_path}")
-    
+
     click.echo(f"Processing images from: {image_path}")
     click.echo(f"Output directory: {output_path}")
     click.echo(f"Model: {model}")
     click.echo(f"Concurrent requests: {concurrent}")
     click.echo(f"Max images: {max_images}")
-    
+
     # Run async function
     results = asyncio.run(auto_predict_images(
         image_dir=image_path,
@@ -145,12 +147,12 @@ def batch_predict(image_dir: str, output_dir: str, model: str, concurrent: int, 
         max_concurrent=concurrent,
         max_images=max_images
     ))
-    
+
     if results:
         completed = sum(1 for r in results if r["status"] == "completed")
         failed = sum(1 for r in results if r["status"] == "failed")
-        
-        click.echo(f"\nSummary:")
+
+        click.echo("\nSummary:")
         click.echo(f"  Total processed: {len(results)}")
         click.echo(f"  Successful: {completed}")
         click.echo(f"  Failed: {failed}")
