--- backend/src/evaluate/__init__.py
@@ -1,1 +1,0 @@
-"""CLI tools for UI annotation evaluation."""
