+++ backend/src/models.py
@@ -1,10 +1,11 @@
 import enum
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, String, DateTime, Enum, Text, Float, Integer, Index
+
+from sqlalchemy import Column, DateTime, Enum, Float, Index, Integer, String, Text
 from sqlalchemy.dialects.postgresql import UUID
 
-from src.db.database import Base
+from src.database.core import Base
 
 
 class JobStatus(str, enum.Enum):
@@ -17,45 +18,45 @@ class JobStatus(str, enum.Enum):
 
 class Job(Base):
     """Job model for tracking image processing tasks."""
-    
+
     __tablename__ = "jobs"
-    
+
     # Primary key
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    
+
     # Job metadata
     status = Column(Enum(JobStatus), default=JobStatus.PENDING, nullable=False, index=True)
     model_name = Column(String(100), nullable=False)
-    
+
     # S3 storage info
     s3_key = Column(String(500), nullable=False)
     s3_url = Column(String(1000))
     original_filename = Column(String(500))
     content_type = Column(String(100))
     file_size = Column(Integer)  # Size in bytes
-    
+
     # Timing information
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False, index=True)
     started_at = Column(DateTime)
     completed_at = Column(DateTime)
     processing_time = Column(Float)  # Time in seconds
-    
+
     # Worker information
     worker_id = Column(String(200))  # Celery task ID
-    
+
     # Results and errors
     result_data = Column(Text)  # JSON string of results
     error_message = Column(Text)
-    
+
     # Client callback (optional)
     callback_url = Column(String(1000))
-    
+
     # Indexes for common queries
     __table_args__ = (
         Index('idx_created_status', 'created_at', 'status'),
         Index('idx_status_created', 'status', 'created_at'),
     )
-    
+
     def to_dict(self):
         """Convert job to dictionary."""
         return {
