+++ backend/src/queue/__init__.py
@@ -1,0 +1,18 @@
+"""Celery task queue components."""
+
+from src.models import Job, JobStatus
+from src.queue.app import celery_app
+from src.queue.tasks import (
+    check_queue_size_task,
+    cleanup_old_jobs_task,
+    process_image_task,
+)
+
+__all__ = [
+    "celery_app",
+    "Job",
+    "JobStatus",
+    "process_image_task",
+    "check_queue_size_task",
+    "cleanup_old_jobs_task"
+]
