+++ backend/tests/test_cli.py
@@ -1,10 +1,11 @@
 """Test the CLI interface."""
 
 import json
+
 import pytest
-from pathlib import Path
 from click.testing import CliRunner
-from src.cli import cli
+
+from src.cli.main import cli
 
 
 @pytest.fixture
@@ -15,7 +16,7 @@ def test_data_dir(tmp_path):
     pred_dir = tmp_path / "predictions"
     gt_dir.mkdir()
     pred_dir.mkdir()
-    
+
     # Create ground truth file
     gt_data = {
         "imageName": "test.png",
@@ -33,10 +34,10 @@ def test_data_dir(tmp_path):
         ],
         "metadata": {"totalAnnotations": 1, "exportedAt": "2025-01-20T10:00:00Z"}
     }
-    
+
     with open(gt_dir / "test_annotations.json", 'w') as f:
         json.dump(gt_data, f)
-    
+
     # Create predictions file
     pred_data = {
         "imageName": "test.png",
@@ -54,23 +55,23 @@ def test_data_dir(tmp_path):
         ],
         "metadata": {"totalAnnotations": 1, "exportedAt": "2025-01-20T10:05:00Z"}
     }
-    
+
     with open(pred_dir / "test_annotations.json", 'w') as f:
         json.dump(pred_data, f)
-    
+
     return tmp_path
 
 
 def test_cli_basic_usage(test_data_dir):
     """Test basic CLI usage."""
     runner = CliRunner()
-    
+
     result = runner.invoke(cli, [
         "evaluate",
         str(test_data_dir / "ground_truth"),
         str(test_data_dir / "predictions")
     ])
-    
+
     assert result.exit_code == 0
     assert "EVALUATION RESULTS" in result.output
     assert "button" in result.output
@@ -82,14 +83,14 @@ def test_cli_basic_usage(test_data_dir):
 def test_cli_with_iou_threshold(test_data_dir):
     """Test CLI with custom IoU threshold."""
     runner = CliRunner()
-    
+
     result = runner.invoke(cli, [
         "evaluate",
         str(test_data_dir / "ground_truth"),
         str(test_data_dir / "predictions"),
         "--iou-threshold", "0.7"
     ])
-    
+
     assert result.exit_code == 0
 
 
@@ -97,22 +98,22 @@ def test_cli_with_output_file(test_data_dir):
     """Test CLI with output file."""
     runner = CliRunner()
     output_file = test_data_dir / "results.json"
-    
+
     result = runner.invoke(cli, [
         "evaluate",
         str(test_data_dir / "ground_truth"),
         str(test_data_dir / "predictions"),
         "--output", str(output_file)
     ])
-    
+
     assert result.exit_code == 0
     assert output_file.exists()
     assert f"Results saved to: {output_file}" in result.output
-    
+
     # Check output file content
     with open(output_file) as f:
         data = json.load(f)
-    
+
     assert "evaluation_params" in data
     assert "per_tag_metrics" in data
     assert "overall_metrics" in data
@@ -122,13 +123,13 @@ def test_cli_with_output_file(test_data_dir):
 def test_cli_invalid_directory():
     """Test CLI with invalid directory."""
     runner = CliRunner()
-    
+
     result = runner.invoke(cli, [
         "evaluate",
         "/nonexistent/path",
         "/another/nonexistent/path"
     ])
-    
+
     assert result.exit_code != 0
     assert "does not exist" in result.output
 
@@ -136,9 +137,9 @@ def test_cli_invalid_directory():
 def test_cli_help():
     """Test CLI help output."""
     runner = CliRunner()
-    
+
     result = runner.invoke(cli, ["evaluate", "--help"])
-    
+
     assert result.exit_code == 0
     assert "Usage: cli evaluate" in result.output
     assert "GROUND_TRUTH_DIR" in result.output
@@ -151,14 +152,14 @@ def test_cli_help():
 def test_cli_with_show_errors(test_data_dir):
     """Test CLI with --show-errors flag."""
     runner = CliRunner()
-    
+
     result = runner.invoke(cli, [
         "evaluate",
         str(test_data_dir / "ground_truth"),
         str(test_data_dir / "predictions"),
         "--show-errors"
     ])
-    
+
     assert result.exit_code == 0
     assert "EVALUATION RESULTS" in result.output
     assert "FP" in result.output  # Should show FP column
