+++ backend/tests/test_metrics.py
@@ -1,7 +1,8 @@
 """Test metrics calculation."""
 
 import pytest
-from src.evaluate.calculate_metrics import calculate_metrics
+
+from src.cli.evaluate.calculate_metrics import calculate_metrics
 
 
 def test_metrics_perfect_predictions():
@@ -48,11 +49,11 @@ def test_metrics_with_real_example():
     """Test with real example from our test data."""
     # Button: TP=3, FP=2, FN=1
     precision, recall, f1 = calculate_metrics(tp=3, fp=2, fn=1)
-    
+
     expected_precision = 3 / (3 + 2)  # = 0.6
     expected_recall = 3 / (3 + 1)     # = 0.75
     expected_f1 = 2 * (0.6 * 0.75) / (0.6 + 0.75)  # ≈ 0.667
-    
+
     assert abs(precision - expected_precision) < 0.001
     assert abs(recall - expected_recall) < 0.001
     assert abs(f1 - expected_f1) < 0.001
@@ -68,10 +69,10 @@ def test_metrics_with_real_example():
 def test_metrics_parametrized(tp, fp, fn, expected_precision, expected_recall):
     """Test metrics with various parameter combinations."""
     precision, recall, f1 = calculate_metrics(tp, fp, fn)
-    
+
     assert abs(precision - expected_precision) < 0.001
     assert abs(recall - expected_recall) < 0.001
-    
+
     # Verify F1 calculation
     if precision + recall > 0:
         expected_f1 = 2 * (precision * recall) / (precision + recall)
